removed sql from incusd
