+++ cmd/incus/admin.go
@@ -19,10 +19,13 @@ func (c *cmdAdmin) Command() *cobra.Command {
 	cmd.Short = i18n.G("Manage incus daemon")
 	cmd.Long = cli.FormatSection(i18n.G("Description"), i18n.G(
 		`Manage incus daemon`))
-
 	// init
 	adminInitCmd := cmdAdminInit{global: c.global}
 	cmd.AddCommand(adminInitCmd.Command())
+
+	// sql
+	adminSqlCmd := cmdAdminSql{global: c.global}
+	cmd.AddCommand(adminSqlCmd.Command())
 
 	return cmd
 }
